% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22084R.R
\name{lasso}
\alias{lasso}
\title{Raw lasso method using optim}
\usage{
lasso(x, y, lambda)
}
\arguments{
\item{x}{the design matrix}

\item{y}{response vector}

\item{lambda}{the punishment coefficient lambda}
}
\value{
the estimator of beta
}
\description{
given punishment coefficient lambda and initial preconceived beta, this function return estimator of beta
}
\examples{
\dontrun{
data(X)
attach(X)
data(Y)
attach(Y)
data(lambda_univ)
attach(lambda_univ)

beta_hat <- lasso(X,Y,lambda_univ)
}
}
