% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22084R.R
\name{LDPE}
\alias{LDPE}
\title{Low dimensional projection estimator}
\usage{
LDPE(x, y)
}
\arguments{
\item{x}{the design matrix}

\item{y}{response vector}
}
\value{
the estimators of beta and sd(sigma) of beta
}
\description{
use LDPE method to give a smaller bias estimator for beta_j
}
\examples{
\dontrun{
data(data)
coef<-LDPE(X,Y)

}
}
