% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22084R.R
\name{RLDPE_z_generate}
\alias{RLDPE_z_generate}
\title{Generate score vector z_j for each column x_j of design X using RLPDE method}
\usage{
RLDPE_z_generate(x, m)
}
\arguments{
\item{x}{the design matrix}

\item{m}{the selected number of maximum correlated beta_i of beta_j}
}
\value{
score vector z
}
\description{
Generate score vector z_j for each column x_j of design X using RLPDE method
}
\examples{
\dontrun{
data(data)
Z<-RLDPE_z_generate(X)
}
}
