% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22084R.R
\name{z_generate}
\alias{z_generate}
\title{Generate score vector z_j for each column x_j of design X}
\usage{
z_generate(x)
}
\arguments{
\item{x}{the design matrix}
}
\value{
score vector z
}
\description{
Generate score vector z_j for each column x_j of design X
}
\examples{
\dontrun{
data(data)
Z<-z_generate(X)

}
}
