% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22084R.R
\name{RLDPE}
\alias{RLDPE}
\title{restricted low dimensional projection estimator}
\usage{
RLDPE(x, y, m)
}
\arguments{
\item{x}{the design matrix}

\item{y}{response vector}

\item{m}{the selected number of maximum correlated beta_i of beta_j}
}
\value{
the estimators of beta and sd(sigma) of beta \code{n}
}
\description{
use RLDPE method to give a smaller bias estimator for beta_j
}
\examples{
\dontrun{
data(data)
m<-4
coef<-RLDPE(X,Y,m)
}
}
