% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22084R.R
\name{scale_lasso}
\alias{scale_lasso}
\title{scale lasso method using optim}
\usage{
scale_lasso(x, y, lambda)
}
\arguments{
\item{x}{the design matrix}

\item{y}{response vector}

\item{lambda}{the punishment coefficient lambda}
}
\value{
the estimator of beta and sigma
}
\description{
given punishment coefficient lambda and initial preconceived beta, this function return estimator of beta
}
\examples{
\dontrun{
data(data)
beta_hat <- scale_lasso(X,Y,lambda_univ)
}
}
